requirements/base.txt
django==4.2.15
python-dotenv==1.0.1
djangorestframework==3.15.2
django-countries==7.6.1
django-phonenumber-field==8.0.0
phonenumbers==8.13.42
drf-spectacular==0.27.2
Pillow==10.4.0
argon2-cffi==23.1.0
djoser==2.3.1
django-filter==24.3
django-celery-email==3.0.0
django-celery-beat==2.6.0
python-dateutil==2.9.0
psycopg2-binary==2.9.9
cloudinary==1.39.1
loguru==0.7.2
redis==5.0.3
celery==5.3.6
flower==2.0.1
django-redis==5.4.0
reportlab==4.4.2